* Depency binaries will be copied to the output directory now ..... TODO: Check if it works well for the "Debug" builds ..... Further notes are described as comment to file ..... "AssembleBinaryFiles" at line ~80
* Assembly info update:
..... * Removed ".*" and the 4th place, because by this a
.....   random(?) number was created (e.g. "1.7.4.22757")
..... * Update of assembly description
.....   Example: "Created by MogreBuilder on 2012-04-15"
..... * Update of "AssemblyConfigurationAttribute" --> Debug/Release

* Added method descriptions for Task.ModifyFile() methods

* Added IO error handling and path check to task "AssembleBinaryFiles"

+++ Source/Tasks/AssembleBinaryFiles.cs
@@ -16,7 +16,6 @@ namespace Mogre.Builder.Tasks
 
         public override void Run()
         {            
-            //String targetDir = String.Format(@"bin\{0}.NET4", inputManager.BuildConfiguration);
             String targetDir = inputManager.BuildOutputDirectory;
 
             if (!Directory.Exists(targetDir))
@@ -27,26 +26,74 @@ namespace Mogre.Builder.Tasks
                 String.Format(@"Main\lib\{0}", inputManager.BuildConfiguration),
                 String.Format(@"Main\OgreSrc\build\bin\{0}", inputManager.BuildConfiguration),
                 String.Format(@"Main\OgreSrc\build\lib\{0}", inputManager.BuildConfiguration),
+                Path.Combine(inputManager.DependenciesDirectory, "bin", inputManager.BuildConfiguration) // Main\OgreSrc\ogre\Dependencies\bin\...
             };
 
             String[] patterns = new String[] { "*.dll", "*.lib", "*.pdb" };
 
             foreach (String binPath in binPaths)
             {
+                // check if path exists
+                if (Directory.Exists(binPath) == false)
+                {
+                    outputManager.Warning("Can't find binary directory:  " + binPath);
+                    continue;
+                }
+
                 foreach (String pattern in patterns)
                 {
                     foreach (String file in Directory.GetFiles(binPath, pattern))
                     {
-                        String filePath = targetDir + "\\" + Path.GetFileName(file);
+                        try
+                        {
+                            String filePath = Path.Combine(targetDir, Path.GetFileName(file));
 
-                        if (File.Exists(filePath))
-                            File.Delete(filePath);
+                            if (File.Exists(filePath))
+                                File.Delete(filePath);
 
-                        File.Move(file, filePath);
+                            File.Move(file, filePath);
+                        }
+                        catch (IOException e)
+                        {
+                            outputManager.Warning(String.Format(
+                                "Can't move binary file '{0}' to output directory. \n" + 
+                                "    Source:     {1}  \n" + 
+                                "    Exception:  {2}" , 
+                                Path.GetFileName(file), Path.GetDirectoryName(file), e.Message));
+                        }
                     }
                 }
             } // foreach
 
+
+            // warning if neither "Debug" nor "Release"
+            if ((inputManager.BuildConfiguration.ToLower() != "debug") &&
+                (inputManager.BuildConfiguration.ToLower() != "release"))
+            {
+                outputManager.Warning(String.Format(
+                    "Note:  The build configuration is not 'Debug' and not 'Release'. (It's '{0}') \n " + 
+                    "       Therefore it's possible, that some binary files are missed in the output directory (e.g. 'cg.dll'). \n" +
+                    "       Consider to search and copy them manually.",
+                    inputManager.BuildConfiguration));
+            }
+
+            // depency warning if "Debug"
+            //   --> TODO: Create an advanced copy task later for Debug builds
+            //       For this we need to check what's the name of the built files (could contain "_d" in the file name) 
+            //       and if they are written to bin\debug or bin\release
+            //       Also the behaviour could be change in newer Ogre versions. (related to the used CMake or VS config files)
+            //       Idea: Copy all files from bin\debug AND bin\release, when the DATE_TIME is newer than MogreBuilder start time.
+            //       For this the 'FileCopyTask' class could be extended and used.
+            if (inputManager.BuildConfiguration.ToLower() == "debug")
+            {
+                outputManager.Warning(String.Format(
+                    "Note:  The build configuration is 'Debug'. \n " +
+                    "       Therefore it's possible, that some binary files (of depencies) are missed in the output directory. \n" +
+                    "       Consider to look to the 'release' binary directory of the depencies. \n" +
+                    "       Path:   '{0}'",
+                    Path.Combine(inputManager.DependenciesDirectory, inputManager.BuildConfiguration)
+                    ));
+            }
 
             outputManager.SuccessfulOgreBuild = true;
